Move functions to modules to improve readability. Update prod dependencies
