+++ lambda/package.json
@@ -1,23 +1,24 @@
 {
-  "name": "serverless-with-lambda-layer",
+  "name": "servicio-de-tizada-lambda",
   "version": "0.0.0",
-  "description": "This package demonstrates using @sparticuz/chromium as a devDependency with a layer that contains the binaries",
+  "description": "Servicio para crear tizadas",
   "license": "ISC",
   "author": {
-    "name": "Kyle McNally"
+    "name": "SmartFactory"
   },
   "main": "index.js",
   "scripts": {
-    "deploy": "test/node_modules/.bin/sls deploy",
-    "test": "test/node_modules/.bin/sls invoke --function chromium-test --log"
+    "test": "jest"
   },
   "dependencies": {
+    "@sparticuz/chromium": "123.0.0",
     "aws-sdk": "^2.1682.0",
     "puppeteer-core": "^19.6.3",
     "winston": "^3.14.2"
   },
   "devDependencies": {
-    "@sparticuz/chromium": "123.0.0",
+    "aws-sdk-mock": "^6.2.0",
+    "jest": "^29.7.0",
     "serverless": "^3.27.0"
   }
 }
