+++ lambda/resources/utils/async_invoke.js
