+++ lambda/resources/utils/s3_upload.js
