+++ lambda/tests/testBuildPartsSVG.js
