+++ lambda/tests/testFetchFromS3.js
