+++ lambda/tests/testUploadToS3.js
